requirements.txt
fastapi
transformers
torch
newspaper3k
lxml_html_clean
nltk
